Slightly more sane nomenclature
With the DSN abstracted away, it is actually easier now to just point dagster to
a different DSN than have different repos for different run environments.
That is why I am dropping the `_test` suffixes and putting a reference to
the repo in the root __init__.py.
I also rename the generic `postgres_connection` to `local_postgres`,
 since there will have to be two more `PostgresConnection` instances in production
 for the remote, longterm data storage.