Bugfix: wrong column in `extract_lines`
This is rather weird: the `extract_lines` query referenced to `line_id`,
when it's primary key is actually just `id`.
I have no clue how that ever worked.
Also, the insert statement was wrapped in a CTE,
that would the just be selected without any modification.
Got rid of that as I do not see the point.